Fix problem with dollar signs being eaten
This occurred when combining the rendered html pages, when the content
was being composed into a single page using penlight's text templating
system. It turns out that two rounds of substitution are done - so that
any words preceded by a dollar sign in the content are treated as
variables, as well as in the template. To compensate for this, we now
double any such dollar signs we come across.